requirements.txt
requests
google-generativeai
gtts
moviepy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2